src-code/app/services/journey_image_generator.py
"""
AI-Powered Patient Journey Image Generator
Uses Google Gemini/Imagen to generate contextual journey visualizations
"""
import os
from dotenv import load_dotenv
import google.generativeai as genai
from PIL import Image
import io
import base64
import requests

load_dotenv()

class JourneyImageGenerator:
    """Generate patient journey visualizations using AI"""
    
    def __init__(self):
        """Initialize Image Generation API"""
        # We can use Gemini, Google, or OpenAI keys
        self.gemini_key = os.getenv("GEMINI_API_KEY") or os.getenv("GOOGLE_API_KEY")
        self.openai_key = os.getenv("OPENAI_API_KEY")
        
        if self.gemini_key:
            genai.configure(api_key=self.gemini_key)
            print("✓ Gemini API configured for journey images")
        
        if not self.gemini_key and not self.openai_key:
            print("⚠️ WARNING: Neither GEMINI_API_KEY nor OPENAI_API_KEY found. Journey image generation will be disabled.")
        
        print("✓ Journey Image Generator initialized")
    
    def generate_journey_images(self, patient_data: dict) -> dict:
        """
        Generate two contextual journey images for a patient
        
        Args:
            patient_data: Patient information including:
                - name: Patient name
                - primary_condition: Main medical condition
                - journey_events: List of timeline events
                - age: Patient age
                - medical_history: Past medical history
                
        Returns:
            dict with image1_url and image2_url (base64 encoded)
        """
        
        name = patient_data.get('name', 'Patient')
        condition = patient_data.get('primary_condition', 'medical condition')
        events = patient_data.get('journey_events', [])
        age = patient_data.get('age', 'adult')
        history = patient_data.get('medical_history', '')
        
        # SANITIZATION: DALL-E 3 safety filters often block "COVID-19" or specific patient names.
        # We replace sensitive terms with clinical equivalents that are safety-filter friendly.
        safe_condition = condition.replace("COVID-19", "Viral Respiratory Condition").replace("Coronavirus", "Viral Infection")
        safe_name = "the patient" # Avoid using real names in image prompts for safety and privacy
        
        # Generate prompts for two different journey aspects using sanitized data
        prompt_1 = self._create_overview_prompt(safe_name, safe_condition, events, age, history)
        prompt_2 = self._create_current_phase_prompt(safe_name, safe_condition, events, age)
        
        print(f"🎨 Generating journey images for {name} ({condition}) - Sanitized for safety")
        
        # Generate images
        try:
            # For demonstration, I'll create a method that generates images
            # You can integrate with:
            # - Google Imagen API
            # - DALL-E API
            # - Stable Diffusion
            # - Or use the generate_image tool
            
            image1_data = self._generate_image_with_api(prompt_1)
            image2_data = self._generate_image_with_api(prompt_2)
            
            return {
                'image1': image1_data,
                'image2': image2_data,
                'prompt1': prompt_1,
                'prompt2': prompt_2
            }
            
        except Exception as e:
            print(f"⚠️ Image generation failed: {e}")
            return {
                'image1': None,
                'image2': None,
                'error': str(e)
            }
    
    def _sanitize_for_ai(self, text: str) -> str:
        """Sanitize text to avoid AI safety filter rejections"""
        if not text: return ""
        # DALL-E 3 is extremely sensitive to COVID-19 and potential violence/gore
        replacements = {
            "COVID-19": "Viral Respiratory Condition",
            "Coronavirus": "Viral Infection",
            "blood": "fluids",
            "crushing": "intense",
            "dying": "critical care",
            "death": "end of journey"
        }
        for key, val in replacements.items():
            text = text.replace(key, val)
        return text

    def _create_overview_prompt(self, name, condition, events, age, history):
        """Create prompt for journey overview visualization"""
        
        # Build event timeline description
        event_desc = ""
        if events:
            recent_events = events[-3:]  # Last 3 events
            event_desc = ". Timeline includes: " + ", then ".join([
                self._sanitize_for_ai(f"{e.get('description', 'event')}") for e in recent_events
            ])
        
        history = self._sanitize_for_ai(history)
        
        prompt = f"""Create a medical journey visualization showing the patient recovery timeline.

Patient Profile:
- Age: {age}
- Condition: {condition}
- Medical History: {history or 'None noted'}

Journey Overview:
{event_desc}

Visual Style:
- Clean, professional medical infographic
- Timeline with key milestones
- Icons representing stages (admission, treatment, recovery, follow-up)
- Soft medical color palette (blues, greens, whites)
- Upward progression showing improvement
- Include graphs showing health metrics improving over time

Format: Horizontal timeline with visual journey map from recovery to current status.
Mood: Hopeful, professional, medical accuracy, patient-centered care.
"""
        return prompt
    
    def _create_current_phase_prompt(self, name, condition, events, age):
        """Create prompt for current recovery phase detail"""
        
        # Get most recent event
        current_phase = "Initial care"
        if events:
            latest = events[-1]
            current_phase = self._sanitize_for_ai(latest.get('description', 'Ongoing treatment'))
        
        prompt = f"""Create a detailed medical visualization of the current treatment phase.

Patient: {age} year old with {condition}
Current Phase: {current_phase}

Visual Requirements:
- Focus on current health status and treatment
- Medical illustrations showing affected area/system
- Treatment approach visualization (medications, therapy, monitoring)
- Progress indicators and vital signs
- Recovery metrics and goals
- Next steps in care plan

Style:
- Medical illustration quality
- Clean, informative design
- Patient education focus
- Anatomical accuracy where relevant
- Positive, encouraging tone

Format: Detailed infographic showing current status, treatment, and recovery roadmap.
Color scheme: Professional medical (white, light blue, green accents).
"""
        return prompt
    
    def _generate_image_with_api(self, prompt: str) -> str:
        """
        Generate image using OpenAI DALL-E 3 (reliable) or Google Imagen
        """
        if not self.openai_key and not self.gemini_key:
            print("❌ Cannot generate image: No API keys configured")
            return None

        # Try OpenAI DALL-E 3 first as we likely have the key
        if self.openai_key:
            try:
                from openai import OpenAI
                client = OpenAI(api_key=self.openai_key)
                
                print("🎨 Generating with DALL-E 3...")
                response = client.images.generate(
                    model="dall-e-3",
                    prompt=prompt[:4000], # Limit length
                    size="1024x1024",
                    quality="standard",
                    n=1,
                    response_format="b64_json" 
                )
                
                b64_data = response.data[0].b64_json
                return f"data:image/png;base64,{b64_data}"
                
            except Exception as e:
                print(f"⚠️ DALL-E generation failed: {e}")
        
        # Fallback to Gemini/Imagen if configured
        if self.gemini_key:
            try:
                # Placeholder for actual Gemini Image Gen call
                # Note: Currently Gemini 1.5 Pro/Flash doesn't natively generate images via this specific SDK
                # usually it's a separate Imagen endpoint or part of Vertex AI.
                print("🎨 Gemini Image generation not yet implemented in this SDK version")
                return None
            except Exception as e:
                print(f"Gemini generation failed: {e}")

        return None

# Singleton instance
journey_image_generator = JourneyImageGenerator()
